requirements.txt
streamlit
pandas
plotly
python-dotenv
requests
fastapi
uvicorn
pytchat